+++ R/geom_helpers.R
@@ -115,7 +115,7 @@ vec_arith.defaulted <- function(op, x, y, ...) {
   vec_arith_base(op, x, y)
 }
 
-as_grouped_variable <- function(x) {
+as_grouped_variable <- function(x = numeric()) {
   vctrs::new_vctr(x, class = "grouped_variable", inherit_base_type = TRUE)
 }
 
@@ -145,3 +145,22 @@ vec_cast.defaulted.double <- function(x, to, ...) set_default(x)
 #' @export
 #' @method vec_cast.double defaulted
 vec_cast.double.defaulted <- function(x, to, ...) as.double(vec_data(x))
+
+#' @export
+vec_ptype2.grouped_variable.double <- function(x, y, ...) as_grouped_variable()
+#' @export
+vec_ptype2.double.grouped_variable <- function(x, y, ...) as_grouped_variable()
+#' @export
+vec_casted.grouped_variable.double <- function(x, to, ...) as_grouped_variable(x)
+#' @export
+#' @method vec_cast.double grouped_variable
+vec_cast.double.grouped_variable <- function(x, to, ...) as.double(vec_data(x))
+#' @export
+vec_arith.grouped_variable <- function(op, x, y, ...) {
+  UseMethod('vec_arith.grouped_variable', y)
+}
+#' @export
+#' @method vec_arith.grouped_variable default
+vec_arith.grouped_variable.default <- function(op, x, y, ...) {
+  as_grouped_variable(vec_math_base(op, vec_data(x), y))
+}