+++ R/geom_rect_theme.R
@@ -41,10 +41,7 @@ geom_rect_theme <- function(mapping = NULL, data = NULL,
                             show.legend = NA,
                             inherit.aes = TRUE,
                             element = NULL) {
-  if (!is.null(element) &&
-      !inherits(element, c("element_rect", "element_blank"))) {
-    stop("The `element` argument should be of type `element_rect`")
-  }
+  element <- .check_geom_element(element, "rect")
   layer(
     data = data,
     mapping = mapping,
@@ -71,10 +68,7 @@ geom_tile_theme <- function(
   linejoin = "mitre", na.rm = FALSE, show.legend = NA, inherit.aes = TRUE,
   element = NULL
 ) {
-  if (!is.null(element) &&
-      !inherits(element, c("element_rect", "element_blank"))) {
-    stop("The `element` argument should be of type `element_rect`")
-  }
+  element <- .check_geom_element(element, "rect")
   layer(
     data = data,
     mapping = mapping,